requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
playwright==1.52.0
readability-lxml==0.8.4.1
lxml[html_clean]==5.3.2
httpx==0.28.1
redis==5.3.0
stripe==12.2.0
python-dotenv==1.1.0
pydantic==2.11.3
slowapi==0.1.9
